SXC When opening a wall next to the player, update the shroud
Tested with a full playthrough for the following maps so far,
further testing will be done with the release candidate.
* Rush
* Mountains Of Doom

I left a few of the maps unchanged, the ones where the change is
never an ambush and the players already have good visibility.
For example, in SXC_Advanced the terrain changes are just
changing a chokepoint in to a 3-hex-wide castle.