SXC Do not allow_undo when a player moves on to the shop
The warning message about a player is shopping sometimes appears too
late, it seems to only show up when they finish shopping. This was
because of the allow_undo in the event, it only gets synchronised by a
later event.

The allow_undo has been there since the first version that stf
included in source control (the Wesnoth 1.6 version), so I'm wary to
remove it, however undoing the first move to the shop causes an OOS,
so it's probably not used.