SXC Pressure add variants menu (Mermaid's Lake / Shallow Water)
This adds a dialog that appears before the SXC difficulty dialog,
allowing the map to be played with low water from the start (which
is expected to make the first 30-40 turns easier, and get the map
explored faster).

I've also added some events that trigger after the variant is
selected, with the idea that this variant support may later become a
feature of SXC that's used by other maps.  In the shallow-water
variant, some chests are added to entice players to the north. The
message that deep-water movement isn't required is only shown on the
deep-water version.

The dialogues about opening sluices haven't changed, but players
will probably understand that.