SXC Incomplete support for arbitrary numbers of AI sides
This requires Wesnoth 1.14. It will seem to work on 1.12.6 until
someone saves and reloads.  Saving and reloading on 1.12.6 will reset
the team names, and all team_name based events will stop working.

Use team names "Heroes" and "Enemies" instead of 1,2,3,4,5 and 6,7,8,9.

Handle a variable number of AI sides in the brutal-and-recruits message