SXC Each map shows its own description
All of them already had their own description, but the description in
SXCmacros.cfg overrode them. Remove that, and also add macros for the
standard texts.

This puts the "this map requires Ageless Era" text first, so that the
whole description changes visibly when scrolling through the list of
maps to choose which game to create.

Also, do not recommend an era, as it would be wrong for Ageless maps.