SXC Remove buggy code related to the fearless trait
This code makes no sense to me, AFAICS if it wasn't buggy then for units with
the fearless trait it would remove that trait, and then replace it again with
a fearless trait.

It's buggy, because if the unit doesn't have fearless then the [store_unit]
doesn't find a unit. On both 1.12.6 and 1.13 the [unstore_unit] complains that
the variable doesn't contain unit data; on 1.13.11 it's an error which stops the
WML processing and therefore causes issue #24 (the conversion of abilities to
megacures etc is skipped).

This causes issue #24, and is one of the problems in issue #22.