SXC Hardcoded support for up to 6 AI sides
It would be good to support an arbitrary number of AI sides, but that would
require a larger refactor of SXC. The parts that are already complete are in
the "incomplete support for arbitrary numbers of sides", whereas the parts in
this commit have simply been updated to fit the number of sides in the WIP
maps with ai10 and possibly ai11.

Fix cashstop acting on turn 1 for ai10 and ai11, and add SXC_INCOME_BONUS_6
for maps with these extra sides.

Fix recruit counting for ai10 and ai11.