SXC Spooky Forest Fix a bug with Pokey, give me anti-shield resistance
The bug being that it's possible to kill the zombies at the start, if you
figure out an exact loophole-exploiting build. By "fix with Pokey", I mean
that, if 8 or more zombies are spawnkilled south-west of the wall in the
south-west corner, the third wave of zombies is replaced with 800hp clones of
the Master of Poking.

With Ageless Era's overpowered characters, I can imagine that a few zombies
could be killed at the start, but killing 8 in normal play seems improbable.

He's now an expendable-leader rather than a hero.  Something was breaking
Pokey's ellipse, and I don't see how it worked before, because the engine sees
that he has an eclipse-hero and canrecruit=yes, so tries to find a hero-leader
ellipse.  Solved by changing his crown and halo to expendable-leader instead.
The canrecruit setting itself is done by SXC, seeing as he's a big_boss.
It's broken for Pokey himself, as well as the clones,