SXC Spooky Forest in revive mode, respawn near a different shop
Implement the feature for moving the respawn point in SXCmacros, and
rename the "revive" and "dead" variables to "sxc_revive_mode" and
"sxc_revive_dead" respectively.

When someone respawns, recalculate the fog and shroud at the start of
player 1's turn.

Don't trigger Spooky Forest's clones-of-Pokey in revive mode. It
didn't happen in the game that prompted this commit, but revive mode
players might kill zombies at the start without working out how to
do the squence-break.