SXC Add warning about selling duplicate weapons
In the rare cases where these warnings are shown, the two descriptions may be
the same too, in which case the warning is the grammatically bad "selling X may
also sell X". The condition is rare enough that I don't want to add additional
logic for that.

This includes a refactor, as implementing the warning required the refactor,
and it didn't seem sense to split them in to separate commits.  The existing
code was using $weapon_j as a variable that was incremented in the first loop
(where it was equal to $weapon_i), but then reused later as the limit for
loops. Renamed the latter usage to $weapon_count, refactored the former usage
to either use $weapon_i or the new [for] and [foreach] loops.

A lot more refactoring could be done, but this change has already taken more
time than I wanted to spend.

Example Ageless Era units to test this:
* AE Yokia - Mandragora (is max level at level 1, but has no specials)
* the Fallen - Enforcer (not max level, different specials on the two weapons)
* the Fallen - Liquidator (max level of the Enforcer)