SXC Refactor the on-recruit code to only include SXC_ENEMY_UPGRADE_WEAPON once
The old version added over 800 lines of preprocessed WML for each occurance,
which added up to over 13000 lines, and that was per-map until the changed to
using a [resource] tag.

It's also a block of code that runs every time the AI recruits, so hopefully
this change may speed up the AI turns a bit.