SXC Refactor mega/minorcures and megaregen start-of-side-turn events
These events trigger at the start of each side's turn, and
update the values based on max HP before the turn-refresh.
Functionality should be the same, just less lines of code.
The two levels of curing are now handled in a single event.

Lots of documentation about [foreach] has been added to the
Wesnoth Wiki, including the behaviour of nested loops.