SXC: Don't put the loyal overlay on regular creeps
In 1.16, the usual TRAIT_LOYAL automatically adds an overlay, however in SXC
that means that all the creeps have it. For SXC we want to identify minibosses
by using the hero marker, however that looks very similar to the loyal marker.

A few instances of IS_HERO get changed to TRAIT_LOYAL_HERO.

The Master of Poking has IS_EXPENDABLE_LEADER, so gets the new no-overlay
version of loyal.