SXC Add a couple of Dunefolk to the SXC Default Low Level Era
The Dune Herbalist is a liminal unit, which makes it different to the
existing units in the low level era. Upgrading a Dune Herbalist to L3
currently gets megacures and megaregen, so it's a bit overpowered.
Should probably be nerfed so that regen+4 becomes minorregen instead.

The Dune Skirmisher is lawful, and similar to the Fencer. The diversion
ability might be interesting against bosses, and having a new unit type
gives a bit of variety anyway.