SXC Sell red potions in every shop
Change the default so that every map allows this. Thanks to CCX for the
suggestion, although I think having every potion available would be too
unbalancing; it would also increase the number of options thrown at new
players.

Will see how it plays, not sure if it'll stay in the long term.