SXC Hide unmoved heroes on Braveheart, Mermaid's Lake, Pursuit and System
Makes it easier to keep a spare side that can be given to a player who joins
later, by making the AI ignore this that spare unit.

Many maps will support a spare side even without this, as the active players
are blocking the route to the starting area. Braveheart's design means that
isn't possible without this feature, while Mermaid's Lake, Pursuit and System
are large maps that branch near the starting area, so this just makes it even
more unlikely that an AI unit will end up there instead of attacking an
active hero.