SXC Rewrite the filters for converting [heals] and [regenerate]
One of AE's main heals+4 abilities is now `id=AE_mag_healing4`,
and in 0.3.3.1 it became neither minorcures nor majorcures. That's
the bug that ccX fixed in 0.3.5.

We're on Wesnoth 1.18, so can use `[experimental_filter_ability]`
instead of needing a hardcoded and outdated list of AE's ids for
healing abilities.

This still has a hardcoded list of AE's ids for regen abilities
that only work on particular terrains or particular times of day.
Such abilities turn into minorregenerates.