SXC Support an arbitrary number of enemy sides
There's still one limitation, for maps that use SXC_INCOME_BONUS to
give gold on each turn when an AI leader has already been slain.
Macros take a fixed number of positional arguments, so
SXC_INCOME_BONUS_10 is limited to bonuses for the first 10 sides.

The implementation underneath deals with a comma-separated list and
can handle arbitrary numbers of sides, but please consider than to be
an internal function which might change in future.